Fix macOS 26 KEXT SDK API compatibility (MMIO/DMA descriptors)
CI on macos-26 (Xcode 26.6) failed because the driver used APIs that
exist only in the local kernel-API stub, not in the public KEXT SDK:

- IORead32/IOWrite32 are not part of the public IOKit kernel KPI.
  Replace with an explicit big-endian MMIO helper on the BAR0 mapping
  (mlx5 init-segment registers are BE32), ordered with OSSynchronizeIO().
- OSMemoryBarrier is hidden behind KERNEL_PRIVATE in OSAtomic.h. Use a
  portable __atomic_thread_fence for DMA ownership ordering and route
  all access through MlxKernelCompat.hpp.
- IOBufferMemoryDescriptor lives in its own header and is returned with
  a 4-argument inTaskWithPhysicalMask; its getBytesNoCopy() is not on
  the base IOMemoryDescriptor. Include the real header, type descriptor
  fields as IOBufferMemoryDescriptor, and drop the bogus 5th argument.
- Fix UAR kernel mapping lifetime: hold the IOMemoryMap instead of
  releasing it after reading the virtual address; derive the BF-register
  address from the live mapping.
- Fix mlx_fw_ver_resp.deviceId to uint16_t (was uint8_t, truncated the
  PCI device ID); struct size is unchanged so the ABI stays intact.

Local stub now mirrors the real SDK shapes (separate IOBufferMemoryDescriptor
header, 4-arg factory, no fabricated MMIO/barrier APIs), so syntax checks
no longer mask these issues. Kernel 20/20 and userspace 15/15 PASS.

+++ Sources/userclient/MlxUCIO.h
@@ -151,7 +151,7 @@ struct mlx_fw_cmd_resp {
 struct mlx_fw_ver_resp {
     uint32_t  fwRev;           /* version number (encoded) */
     uint32_t  cmdifRev;
-    uint8_t   deviceId;
+    uint16_t  deviceId;
     uint8_t   portType;
     uint32_t  numPorts;
 };
